Split CompilerGeneratedMetadataManager
Two classes now:

* Things that are shared by all MD managers that generate metadata
* Things that are only relevent for MD managers that generate statical-usage-based metadata
