+++ src/ILCompiler.Compiler/src/Compiler/UsageBasedMetadataManager.cs
@@ -1,0 +1,143 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+using Internal.TypeSystem;
+
+using ILCompiler.Metadata;
+using ILCompiler.DependencyAnalysis;
+
+using DependencyList = ILCompiler.DependencyAnalysisFramework.DependencyNodeCore<ILCompiler.DependencyAnalysis.NodeFactory>.DependencyList;
+
+namespace ILCompiler
+{
+    /// <summary>
+    /// This class is responsible for managing native metadata to be emitted into the compiled
+    /// module. It applies a policy that every type/method that is statically used shall be reflectable.
+    /// </summary>
+    public sealed class UsageBasedMetadataManager : GeneratingMetadataManager
+    {
+        public UsageBasedMetadataManager(
+            CompilationModuleGroup group,
+            CompilerTypeSystemContext typeSystemContext,
+            MetadataBlockingPolicy blockingPolicy,
+            string logFile,
+            StackTraceEmissionPolicy stackTracePolicy)
+            : base(group, typeSystemContext, blockingPolicy, logFile, stackTracePolicy)
+        {
+        }
+
+        protected override MetadataCategory GetMetadataCategory(FieldDesc field)
+        {
+            MetadataCategory category = 0;
+
+            if (!IsReflectionBlocked(field))
+            {
+                category = MetadataCategory.RuntimeMapping;
+
+                if (_compilationModuleGroup.ContainsType(field.GetTypicalFieldDefinition().OwningType))
+                    category |= MetadataCategory.Description;
+            }
+
+            return category;
+        }
+
+        protected override MetadataCategory GetMetadataCategory(MethodDesc method)
+        {
+            MetadataCategory category = 0;
+
+            if (!IsReflectionBlocked(method))
+            {
+                category = MetadataCategory.RuntimeMapping;
+
+                if (_compilationModuleGroup.ContainsType(method.GetTypicalMethodDefinition().OwningType))
+                    category |= MetadataCategory.Description;
+            }
+
+            return category;
+        }
+
+        protected override MetadataCategory GetMetadataCategory(TypeDesc type)
+        {
+            MetadataCategory category = 0;
+
+            if (!IsReflectionBlocked(type))
+            {
+                category = MetadataCategory.RuntimeMapping;
+
+                if (_compilationModuleGroup.ContainsType(type.GetTypeDefinition()))
+                    category |= MetadataCategory.Description;
+            }
+
+            return category;
+        }
+
+        protected override void ComputeMetadata(NodeFactory factory,
+            out byte[] metadataBlob,
+            out List<MetadataMapping<MetadataType>> typeMappings,
+            out List<MetadataMapping<MethodDesc>> methodMappings,
+            out List<MetadataMapping<FieldDesc>> fieldMappings,
+            out List<MetadataMapping<MethodDesc>> stackTraceMapping)
+        {
+            ComputeMetadata(new GeneratedTypesAndCodeMetadataPolicy(_blockingPolicy, factory),
+                factory, out metadataBlob, out typeMappings, out methodMappings, out fieldMappings, out stackTraceMapping);
+        }
+
+        protected override void GetMetadataDependenciesDueToReflectability(ref DependencyList dependencies, NodeFactory factory, MethodDesc method)
+        {
+            dependencies = dependencies ?? new DependencyList();
+            dependencies.Add(factory.MethodMetadata(method.GetTypicalMethodDefinition()), "Reflectable method");
+        }
+
+        protected override void GetMetadataDependenciesDueToReflectability(ref DependencyList dependencies, NodeFactory factory, TypeDesc type)
+        {
+            TypeMetadataNode.GetMetadataDependencies(ref dependencies, factory, type, "Reflectable type");
+        }
+
+        private struct GeneratedTypesAndCodeMetadataPolicy : IMetadataPolicy
+        {
+            private readonly MetadataBlockingPolicy _blockingPolicy;
+            private readonly NodeFactory _factory;
+            private readonly ExplicitScopeAssemblyPolicyMixin _explicitScopeMixin;
+
+            public GeneratedTypesAndCodeMetadataPolicy(MetadataBlockingPolicy blockingPolicy, NodeFactory factory)
+            {
+                _blockingPolicy = blockingPolicy;
+                _factory = factory;
+                _explicitScopeMixin = new ExplicitScopeAssemblyPolicyMixin();
+            }
+
+            public bool GeneratesMetadata(FieldDesc fieldDef)
+            {
+                return _factory.FieldMetadata(fieldDef).Marked;
+            }
+
+            public bool GeneratesMetadata(MethodDesc methodDef)
+            {
+                return _factory.MethodMetadata(methodDef).Marked;
+            }
+
+            public bool GeneratesMetadata(MetadataType typeDef)
+            {
+                return _factory.TypeMetadata(typeDef).Marked;
+            }
+
+            public bool IsBlocked(MetadataType typeDef)
+            {
+                return _blockingPolicy.IsBlocked(typeDef);
+            }
+
+            public bool IsBlocked(MethodDesc methodDef)
+            {
+                return _blockingPolicy.IsBlocked(methodDef);
+            }
+
+            public ModuleDesc GetModuleOfType(MetadataType typeDef)
+            {
+                return _explicitScopeMixin.GetModuleOfType(typeDef);
+            }
+        }
+    }
+}
